support for cloudformation outputs
- refactor resource class into base/subclasses
- add output resource subclass
- consolidate collection logic in single class
- misc other updates to accommodate outputs
