+++ cirrus/cli/components/__init__.py
@@ -1,4 +1,4 @@
-from .base import Component
+from .base import Component, Lambda, StepFunction
 from .core import CoreFunction
 from .feeders import Feeder
 from .tasks import Task
