+++ cirrus/cli/resources/outputs/__init__.py
@@ -1,0 +1,5 @@
+from ..base import BaseResource
+
+class Output(BaseResource):
+    top_level_key = 'Outputs'
+    task_batch_resource_attr = 'batch_outputs'
