+++ cirrus/cli/resources/resources/config/basic-compute-environments.yml
