+++ cirrus/cli/resources/resources/config/custom-compute-environments.yml
