+++ cirrus/cli/resources/resources/config/resources.yml
